requirements.txt
black==25.1.0
blinker==1.9.0
build==1.2.2.post1
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy-extensions==1.0.0
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pathspec==0.12.1
pip-review==1.3.0
pip-tools==7.4.1
platformdirs==4.3.7
pluggy==1.5.0
pyproject_hooks==1.2.0
pypylon==4.1.0
pytest==8.3.5
python-dotenv==1.1.0
setuptools==78.1.0
Werkzeug==3.1.3
wheel==0.45.1
